+++ deepsights/contentstore/resources/news.py
@@ -43,11 +43,12 @@ class NewsResource(APIResource):
     """
 
     #################################################
-    def _vector_search(
+    def vector_search(
         self,
         query_embedding: List,
         min_score: float = 0.7,
         max_results: int = 30,
+        languages: List[str] = None,
         recency_weight: float = None,
     ):
         """
@@ -58,6 +59,7 @@ class NewsResource(APIResource):
             query_embedding (List): The embedding vector representing the query.
             min_score (float, optional): The minimum score threshold for search results. Defaults to 0.7.
             max_results (int, optional): The maximum number of search results to return. Defaults to 30.
+            languages (List[str], optional): The list of languages to search for. Defaults to None.
             recency_weight (float, optional): The weight to apply to recency in the search ranking. Defaults to None.
 
         Returns:
@@ -75,13 +77,16 @@ class NewsResource(APIResource):
             min_score=min_score,
             max_results=max_results,
             recency_weight=recency_weight,
+            languages=languages,
         )
 
     #################################################
-    def _text_search(
+    def text_search(
         self,
         query: str,
         max_results: int = 30,
+        offset: int = 0,
+        languages: List[str] = None,
         recency_weight: float = None,
     ):
         """
@@ -91,6 +96,8 @@ class NewsResource(APIResource):
 
             query (str): The search query.
             max_results (int, optional): The maximum number of search results to return. Defaults to 30.
+            offset (int, optional): The offset to start the search from. Defaults to 0.
+            languages (List[str], optional): The list of languages to search for. Defaults to None.
             recency_weight (float, optional): The weight to assign to recency in the search ranking. Defaults to None.
 
         Returns:
@@ -106,6 +113,8 @@ class NewsResource(APIResource):
             query=query,
             max_results=max_results,
             recency_weight=recency_weight,
+            languages=languages,
+            offset=offset,
         )
 
     #################################################
@@ -113,6 +122,7 @@ class NewsResource(APIResource):
         self,
         query: str,
         max_results: int = 30,
+        languages: List[str] = None,
         min_vector_score: float = 0.7,
         vector_fraction: float = 0.9,
         vector_weight: float = 0.9,
@@ -129,6 +139,7 @@ class NewsResource(APIResource):
             query (str): The query.
             search_result (BaseModel): The model to use for parsing search results.
             max_results (int, optional): The maximum number of search results to return. Defaults to 30.
+            languages (List[str], optional): The list of languages to search for. Defaults to None.
             min_vector_score (float, optional): The minimum score threshold for search results. Defaults to 0.7.
             vector_fraction (float, optional): The fraction of the search results to be vector-based. Defaults to 0.9.
             vector_weight (float, optional): The weight to apply to vector search in result ranking. Defaults to 0.9.
@@ -156,4 +167,5 @@ class NewsResource(APIResource):
             promote_exact_match=promote_exact_match,
             search_from_timestamp=search_from_timestamp,
             search_to_timestamp=search_to_timestamp,
+            languages=languages,
         )